linux-toradex: update to Toradex V2.5 Beta 2 release
Update git revision to match the latest Toradex BSP release V2.5
Beta 2. The kernel is based on the 4.1.12 release of the upstream
kernel with some new updates and fixes. The NAND flash controller
driver has been synchronized with the latest upstream version,
which also needed a incompatible device tree change. Besides that,
automatic USB host/device change for the Colibri USBC-Port is
supported now and the UART's DMA mode has been disabled by default
due to instability issues. Refer to the Toradex release notes for
the full list of changes.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>